\name{GGEBiplot}
\Rdversion{1.0}
\alias{GGEBiplot}
\title{
Interactive GGE Biplots in R
}
\description{
Provides a graphical user interface for the construction of, interaction with, and manipulation of GGE biplots in R.
}
\usage{
GGEBiplot()
}
\details{
Some of the functions are: 
\cr (i) ranking the cultivars based on their performance in any given environment, 
\cr(ii) ranking the environments based on the relative performance of any given cultivar, 
\cr (iii) comparing the performance of any pair of cultivars in diferent environments, 
\cr (iv) identifying the best cultivar in each environment,
\cr (v) grouping the environments based on the best cultivars, 
\cr (vi) evaluating the cultivars based on both average yield and stability and 
\cr (vii) evaluating the environments based on both discriminating ability and representativeness. 
\cr Three-dimensional biplots are incorporated via the rgl package.
\cr At present, BiplotGUI is intended to be run under Windows.
}
\references{
Yan W, Hunt LA, Sheng Q, Szlavnics Z (2000). "Cultivar evaluation and mega-environment investigation based on GGE biplot." Crop Sci, 40, 597-605.
\cr Yan W, Kang M (2003). GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists, and Agronomists. CRC Press, Boca Raton, FL, USA.
}
\author{
Elisa Frutos Bernal
}
