\name{GGEBiplotGUI-package}
\Rdversion{1.0}
\alias{GGEBiplotGUI-package}
\alias{GGEBiplotGUI}
\docType{package}
\title{
GGEBiplotGUI: Interactive GGE Biplots in R
}
\description{
A GUI with which to construct and interact with GGE biplots.
}
\details{
\tabular{ll}{
Package: \tab GGEBiplotGUI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-27\cr
License: \tab GPL-2 or later\cr
LazyLoad: \tab yes\cr
}
The GGEBiplot function initialises the GUI.
\cr At present, GGEBiplotGUI is intended to be run under Windows. 
}  
\author{
Elisa Frutos Bernal
\cr Maintainer: Who to complain to <efb@usal.es>
}
\references{
Yan W, Hunt LA, Sheng Q, Szlavnics Z (2000). "Cultivar evaluation and mega-environment investigation based on GGE biplot." Crop Sci, 40, 597-605.
\cr Yan W, Kang M (2003). GGE Biplot Analysis: A Graphical Tool for Breeders, Geneticists, and Agronomists. CRC Press, Boca Raton, FL, USA.
}

